Add files via upload
The Indian Premier League is a professional Twenty20 cricket league in India contested during March or April and May of every year by eight teams representing eight different cities or states in India. The league was founded by the Board of Control for Cricket in India in 2008. This is an analysis of all the matches from IPL season 2008 to 2019